+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `/report-token-use` local session token report command.
+
 ## [3.19.5] - 2026-05-26
 
 ### Fixed